rename auto-update-cli to auto-update-check-cli
